ts file refactoring
modular approach
html template files
