+++ assets/scripts/ts/app/app.component.html
@@ -1,0 +1,2 @@
+<header></header>
+<router-outlet></router-outlet>
