+++ assets/scripts/ts/blog/posts.component.html
@@ -1,0 +1,9 @@
+<h2>Posts</h2>
+<ul *ngIf="posts">
+  <li *ngFor="let post of posts">
+    <h1 (click)="showPost(post)">{{ post.title.rendered }}</h1>
+  </li>
+</ul>
+<div *ngIf="currentPost">
+  <p [innerHTML]=currentPost.content.rendered></p>
+</div>
