+++ assets/scripts/ts/custom-posts/custom-posts.component.html
@@ -1,0 +1,6 @@
+<h2>Custom Posts</h2>
+<ul *ngIf="posts">
+  <li *ngFor="let post of posts">
+    <h1>{{ post.title.rendered }}</h1>
+  </li>
+</ul>
