+++ assets/scripts/ts/custom-posts/custom-posts.component.ts
@@ -3,14 +3,7 @@ import { CustomPostsService } from './custom-posts.service';
 
 @Component({
   selector: 'custom-posts',
-  template: `
-    <h2>Custom Posts</h2>
-    <ul *ngIf="posts">
-      <li *ngFor="let post of posts">
-        <h1>{{ post.title.rendered }}</h1>
-      </li>
-    </ul>
-  `,
+  templateUrl: './custom-posts.component.html',
   providers: [CustomPostsService]
 })
 export class CustomPostsComponent implements OnInit {
