+++ assets/scripts/ts/custom-posts/custom-posts.service.ts
