+++ assets/scripts/ts/front-page/front-page.component.html
@@ -1,0 +1,1 @@
+<h2>Home</h2>
