+++ assets/scripts/ts/front-page/front-page.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'home',
+  templateUrl: './front-page.component.html'
+})
+export class FrontPageComponent { }
