+++ assets/scripts/ts/header/header.component.html
@@ -1,0 +1,2 @@
+<h1>Page Title</h1>
+<menu></menu>
