--- assets/scripts/ts/home.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'home',
-  template: `
-    <h2>Home</h2>
-  `,
-})
-export class HomeComponent { }
