+++ assets/scripts/ts/main.ts
@@ -2,7 +2,7 @@
 
 import { bootstrap }    from '@angular/platform-browser-dynamic';
 import { HTTP_PROVIDERS } from '@angular/http';
-import { appRouterProviders } from './app.routes';
-import { AppComponent } from './app.component';
+import { appRouterProviders } from './app/app.routes';
+import { AppComponent } from './app/app.component';
 
 bootstrap(AppComponent, [HTTP_PROVIDERS, appRouterProviders]);
