+++ assets/scripts/ts/menu/menu.component.html
@@ -1,0 +1,6 @@
+<ul>
+  <a [routerLink]="['']">Home</a>
+  <a [routerLink]="['/posts']">Posts</a>
+  <a [routerLink]="['/custom-posts']">Custom Posts</a>
+  <a [routerLink]="['/sample-page']">Sample Page</a>
+</ul>
