+++ assets/scripts/ts/menu/menu.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { ROUTER_DIRECTIVES } from '@angular/router';
+
+@Component({
+  selector: 'menu',
+  templateUrl: './menu.component.html',
+  directives: [ROUTER_DIRECTIVES]
+})
+export class MenuComponent { }
